project2/requirements.txt
grpcio>=1.62.0
grpcio-tools>=1.62.0
protobuf>=4.25.0
numpy>=1.26.0
docker>=7.0
